# Use fuzzy search in a search input field