--- a/frontend/SearchUtils.js
+++ b/frontend/SearchUtils.js
@@ -53,7 +53,68 @@ function trimSearchText(needle: string): string {
   return needle;
 }
 
+function fuzzySearch(needle: ?string, haystack: ?string): Object {
+  var output = {
+    result: false,
+    matches: [],
+  };
+
+  if (!needle || !haystack) {
+    return output;
+  }
+
+  if (needle.length > haystack.length) {
+    return output;
+  }
+
+  if (needle.length === haystack.length) {
+    output.result = needle.toLowerCase() === haystack.toLowerCase();
+
+    if (output.result) {
+      output.matches = needle.split('').map(function(item, itemIndex) {
+        return itemIndex;
+      });
+    }
+
+    return output;
+  }
+
+  var needleIdx = 0;
+  var haystackIdx = 0;
+
+  while ((needleIdx < needle.length) && (haystackIdx < haystack.length)) {
+    var needleChar = needle.charAt(needleIdx).toLowerCase();
+    var haystackChar = haystack.charAt(haystackIdx).toLowerCase();
+
+    if (needleChar === haystackChar) {
+      output.matches.push(haystackIdx);
+
+      needleIdx++;
+    }
+
+    haystackIdx++;
+  }
+
+  if (needleIdx === needle.length) {
+    output.result = true;
+  }
+
+  output.matches = output.result ? output.matches : [];
+
+  return output;
+}
+
+function isFuzzyMatch(needle: string, haystack: string): boolean {
+  return fuzzySearch(needle, haystack).result;
+}
+
+function getFuzzyMatches(needle: string, haystack: string): Array<number> {
+  return fuzzySearch(needle, haystack).matches;
+}
+
 module.exports = {
+  getFuzzyMatches,
+  isFuzzyMatch,
   isValidRegex,
   searchTextToRegExp,
   shouldSearchUseRegex,