Component: Search input in `SettingsPane`.
Proposal: allow searching nodes by partial matches.

Current behaviour:
Input: `Higher`. Output: `<HigherOrderComponent> list`
Input: `hoc`. Output: `no results`
Input: `hdeon`. Output: `no results`

Proposed behaviour:
Input: `Higher`. Output: `<HigherOrderComponent> list`
Input: `hoc`. Output: `<HigherOrderComponent> list`
Input: `hdeon`. Output: `<HigherOrderComponent> list`

Changes results highlighting:
`hoc` => <**H**igher**O**rder**C**omponent>

This fuzzy search algo uses simple match approach (no weights for matches applied).

My questions:
1. Should I replace all `this.searchRoots` filtering with `nodeMatchesText`? See my comment in a code section.
2. Is it okay to output `<span>` for every char in a `Node.name`?
3. `flow check` outputs error about code I didn't touch. What should I do about it?