@bvaughn 2017-05-03
I have some UX concerns with the highlighting behavior in this branch.

In this case, I would expect all 5 of the "a"s to be highlighted, rather than just the first one:
![screen shot 2017-05-03 at 3 51 12 pm](https://cloud.githubusercontent.com/assets/29597/25684827/5cd8e82e-3018-11e7-886b-452f30f32768.png)

In this case, I would expect "AnimatedComp" to be highlighted, since that's the full fuzzy-search match:
![screen shot 2017-05-03 at 3 52 32 pm](https://cloud.githubusercontent.com/assets/29597/25684867/955ff3ea-3018-11e7-9400-09b939eecf7e.png)

Compare to the current highlighting behavior of master:
![screen shot 2017-05-03 at 3 53 34 pm](https://cloud.githubusercontent.com/assets/29597/25684891/b1e4eaac-3018-11e7-857a-dcf46852efdb.png)
![screen shot 2017-05-03 at 3 54 56 pm](https://cloud.githubusercontent.com/assets/29597/25684928/e2d75d70-3018-11e7-8d0a-fda02d8ff098.png)