@olegafx 2017-05-03
@bvaughn my solution was inspired by Sublime Text's file search. Here is how it looks in Sublime:

First case with `a`. I think that it's better to show only first `a` because: 1) highlighting all `a`s is confusing 2) it's better for performance (exit from a search function as soon as possible)
<img width="603" alt="__code_sublime-find-test_animatedcomponent_ _sublime-find-test" src="https://cloud.githubusercontent.com/assets/867546/25685545/f2efa6d8-3070-11e7-822f-c9e6d22a9c19.png">

I find it very useful to see **why** I got these search results. Highlighting only found symbols really helps me to understand it.
<img width="610" alt="__code_sublime-find-test_animatedcomponent_ _sublime-find-test 2" src="https://cloud.githubusercontent.com/assets/867546/25685549/f6e4f806-3070-11e7-8208-a9294e044590.png">