@bvaughn 2017-05-03
The biggest perf-limiting factor at the moment may be that we aren't windowing the nodes we render- but it looks like the new fuzzy search is a bit slower than what's in master. I added some timing logging to the `changeSearch` method. Comparing master to this branch, searching for single letters (so the result set should be the same), average search time went up from ~11.89ms to ~15.04ms for a medium/small test application and from ~17.88ms to ~25.39ms for a larger application.