@bvaughn 2017-05-03
Those are reasonable arguments for the UI behavior in this branch. 👍