@olegafx 2017-05-04
@bvaughn what do you mean by "windowing the nodes"?

It's bad that we should call our function twice – 1st time for a nodes filtering, 2nd time for a name highlighting.

And about a 1 symbol entered. Maybe we should use `indexOf` for that case?