@bvaughn 2017-05-04
> What do you mean by "windowing the nodes"?

Ah, sorry. I meant limiting the number of `Node` elements we render. There's no need to render _all_ of them so long as we render enough to fill the viewable portion of devtools.

> It's bad that we should call our function twice – 1st time for a nodes filtering, 2nd time for a name highlighting.

We call it once to _filter_ nodes (true/false match) and another time to _highlight_ text (splitting it into pieces and mapping them to styles). This seems reasonable to me since those 2 are different operations. Windowing would limit the number of times we made the second call.

> And about a 1 symbol entered. Maybe we should use indexOf for that case?

I wouldn't expect that to make _much_ of a difference vs your current `needleIdx < needle.length` check?