@bvaughn 2017-05-13
Last round of back and forth was me saying,

> Since we're gathering matches both times, we are doing unnecessary work. Maybe we should split that up to avoid building a matches array we aren't using during the filter phase? Or maybe we should attach match metadata to filtered nodes? 

Was kind of waiting to hear back from you and what you thought about that. Also I got distracted by a rather big refactor of the UI/styles for theming support.