@olegafx 2017-05-16
@bvaughn I think that 4–8ms diff is acceptable. In my opinion, the search operation is not frequently used. If someone will notice the difference, we will return to perf optimizations.