@bvaughn 2017-05-20
I'm still a little unsure how I feel about this change, given the performance impact and the fact that devtools was slow to begin with in this area. I'd like to think on it a bit longer and give others a chance to comment if they have any feelings one way or another.

Didn't want you to think that I'd forgotten about it. 😄