@bvaughn 2017-08-30
No harm in leaving it open. I never decided how I felt about it and I've been pulled to work on other projects for the time being. Sorry to leave you hanging!