@bvaughn 2019-08-20
React DevTools has been rewritten and [recently launched a new version 4 UI](https://reactjs.org/blog/2019/08/15/new-react-devtools.html) 🎉 The source code for this rewrite was done in a separate repository and is [being merged into the main React repo](https://github.com/facebook/react/pull/16381).

**Thank you for taking the time to contribute to this extension** 🙇‍♂️

Because this PR is for the version code base, I am closing it. If you'd like to contribute to the new extension, please visit [the new repository](https://github.com/facebook/react/issues?q=is%3Aissue+is%3Aopen+label%3A%22Component%3A+Developer+Tools%22).